# aspnetcore/src/IO.Swagger/Controllers/DashboardApi.cs
/*
 * API Documentation
 *
 * defaultDeription
 *
 * OpenAPI spec version: 1.0.0-oas3
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DashboardApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>Auto generated using Swagger Inspector</remarks>
        /// <param name="accessToken"></param>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("/8API/List/CreditManagement/DashboardDefinition/Description")]
        [ValidateModelState]
        [SwaggerOperation("8APIListCreditManagementDashboardDefinitionDescriptionGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(InlineResponse20013), description: "successful operation")]
        public virtual IActionResult 8APIListCreditManagementDashboardDefinitionDescriptionGet([FromHeader]string accessToken)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(InlineResponse20013));
            string exampleJson = null;
            exampleJson = "{\n  \"result\" : 0,\n  \"data\" : [ {\n    \"Active\" : \"Active\",\n    \"FontHeader\" : \"FontHeader\",\n    \"FontFooter\" : \"FontFooter\",\n    \"Description\" : \"Description\",\n    \"SetAsDefault\" : \"SetAsDefault\",\n    \"UserOrRoleId\" : \"UserOrRoleId\",\n    \"UserId\" : \"UserId\",\n    \"DashboardId\" : \"DashboardId\",\n    \"Default\" : \"Default\",\n    \"Name\" : \"Name\"\n  }, {\n    \"Active\" : \"Active\",\n    \"FontHeader\" : \"FontHeader\",\n    \"FontFooter\" : \"FontFooter\",\n    \"Description\" : \"Description\",\n    \"SetAsDefault\" : \"SetAsDefault\",\n    \"UserOrRoleId\" : \"UserOrRoleId\",\n    \"UserId\" : \"UserId\",\n    \"DashboardId\" : \"DashboardId\",\n    \"Default\" : \"Default\",\n    \"Name\" : \"Name\"\n  } ],\n  \"message\" : \"message\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<InlineResponse20013>(exampleJson)
                        : default(InlineResponse20013);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}