# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "description" : "defaultDeription",
    "version" : "1.0.0-oas3"
  },
  "servers" : [ {
    "url" : "https://0812dce4.ngrok.io"
  } ],
  "paths" : {
    "/8APi/Userlogin" : {
      "post" : {
        "tags" : [ "Login" ],
        "description" : "Login the user  to get the access token. The returned Access token should be pasased in the header in all the further API calls  with the key “AccessToken” as header parameter.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body"
              },
              "examples" : {
                "0" : {
                  "value" : "{\n  \"data\": [\n    {\n      \"Username\":\"su\",\n      \"password\": \"su\",\n      \"CompanyId\": 72,\n      \"DeveloperId\" : \"H3jLzl3ilZU=\"\n    }\n  ],\n  \"result\": 1,\n  \"message\": \"\"\n}\n"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                },
                "examples" : { }
              }
            }
          },
          "403" : {
            "description" : "Forbidden - Access to the resource by the provider is forbidden"
          },
          "404" : {
            "description" : "Not found - The requested resource is not found"
          },
          "502" : {
            "description" : "Provider server error - Something went wrong on the Provider or Endpoint’s server"
          }
        },
        "servers" : [ {
          "url" : "https://0812dce4.ngrok.io"
        } ]
      },
      "servers" : [ {
        "url" : "https://0812dce4.ngrok.io"
      } ]
    },
...
    "/8API/List/CreditManagement/DashboardDefinition/Description" : {
      "get" : {
        "tags" : [ "Dashboard" ],
        "description" : "Auto generated using Swagger Inspector",
        "parameters" : [ {
          "name" : "AccessToken",
          "in" : "header",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          },
          "example" : "hnnju2JcBL9tQd0kmLeLKKbJCX+7Hufv7+wu9TmcTV0="
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_13"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost"
      } ]
    },
...
  },
  "components" : {
    "schemas" : {
      "8APiUserlogin_data" : {
        "type" : "object",
        "properties" : {
          "password" : {
            "type" : "string"
          },
          "CompanyId" : {
            "type" : "integer"
          },
          "Username" : {
            "type" : "string"
          },
          "DeveloperId" : {
            "type" : "string"
          }
        }
      },
      "body" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/8APiUserlogin_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "inline_response_200_data" : {
        "type" : "object",
        "properties" : {
          "AccessToken" : {
            "type" : "string"
          }
        },
        "example" : {
          "AccessToken" : "AccessToken"
        }
      },
      "inline_response_200" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "AccessToken" : "AccessToken"
          }, {
            "AccessToken" : "AccessToken"
          } ],
          "message" : "message"
        }
      },
...
      "inline_response_200_13_data" : {
        "type" : "object",
        "properties" : {
          "Active" : {
            "type" : "string"
          },
          "Description" : {
            "type" : "string"
          },
          "Name" : {
            "type" : "string"
          },
          "UserOrRoleId" : {
            "type" : "string"
          },
          "UserId" : {
            "type" : "string"
          },
          "Default" : {
            "type" : "string"
          },
          "FontHeader" : {
            "type" : "string"
          },
          "DashboardId" : {
            "type" : "string"
          },
          "SetAsDefault" : {
            "type" : "string"
          },
          "FontFooter" : {
            "type" : "string"
          }
        },
        "example" : {
          "Active" : "Active",
          "FontHeader" : "FontHeader",
          "FontFooter" : "FontFooter",
          "Description" : "Description",
          "SetAsDefault" : "SetAsDefault",
          "UserOrRoleId" : "UserOrRoleId",
          "UserId" : "UserId",
          "DashboardId" : "DashboardId",
          "Default" : "Default",
          "Name" : "Name"
        }
      },
      "inline_response_200_13" : {
        "type" : "object",
        "properties" : {
          "result" : {
            "type" : "integer"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/inline_response_200_13_data"
            }
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "result" : 0,
          "data" : [ {
            "Active" : "Active",
            "FontHeader" : "FontHeader",
            "FontFooter" : "FontFooter",
            "Description" : "Description",
            "SetAsDefault" : "SetAsDefault",
            "UserOrRoleId" : "UserOrRoleId",
            "UserId" : "UserId",
            "DashboardId" : "DashboardId",
            "Default" : "Default",
            "Name" : "Name"
          }, {
            "Active" : "Active",
            "FontHeader" : "FontHeader",
            "FontFooter" : "FontFooter",
            "Description" : "Description",
            "SetAsDefault" : "SetAsDefault",
            "UserOrRoleId" : "UserOrRoleId",
            "UserId" : "UserId",
            "DashboardId" : "DashboardId",
            "Default" : "Default",
            "Name" : "Name"
          } ],
          "message" : "message"
        }
      },
...
    }
  }
}